refactor daemon compile pipeline
